PR-11: extract backends/gnome.rs and backends/kde/
Behaviour-preserving relocation. ~800 LOC across 4 files.
- backends/gnome.rs: run_gnome + query_gnome_focus + GNOME focus
  signal subscription helpers.
- backends/kde/mod.rs: KwinScriptGuard + run_kde, plus pub(crate)
  re-exports of script and probe submodules (per D24).
- backends/kde/script.rs: KWin script-generation helpers
  (kwin_*_script_path, KdeFocusQueryService with zbus interface,
  build_kde_query_script, build_kde_focus_push_script).
- backends/kde/probe.rs: KWin runtime-mode probe + script unload
  helpers + query_kde_focus.
- backends/mod.rs: added query_focus_for_env and apply_focus_for_env
  dispatch helpers (the natural home per plan §3 line 113).

main.rs: 3300 -> 2514 LOC. supervisor/mod.rs's run_* imports
updated to crate::backends::{gnome,kde}::run_{gnome,kde}.
pause.rs's apply_focus_for_env import updated to
crate::backends::apply_focus_for_env. Visibility widenings
on these items from PR-06/PR-08 are now redundant (functions
moved); no main.rs widening to revert because the items left
main.rs entirely.

Verification: cargo build clean; cargo test --bin kanata-switcher 261/0.
