+++ src/daemon/main.rs
@@ -66,7 +66,7 @@ use backends::wayland::*;
 
 #[cfg(test)]
 #[allow(unused_imports)]
-pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*, broadcasters::*, kanata::*, control::*, control::client::*, pause::*, focus_pipeline::*, lifecycle::*, lifecycle::logind::*, lifecycle::startup::*, display_override::*, supervisor::*, supervisor::capabilities::*, backends::*, backends::wayland::*};
+pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*, broadcasters::*, kanata::*, control::*, control::client::*, pause::*, focus_pipeline::*, lifecycle::*, lifecycle::logind::*, lifecycle::startup::*, display_override::*, supervisor::*, supervisor::capabilities::*, backends::*, backends::gnome::*, backends::kde::*, backends::kde::script::*, backends::kde::probe::*, backends::wayland::*};
 
 
 // === SNI Indicator ===
@@ -711,270 +711,6 @@ impl Tray for SniIndicator {
     }
 }
 
-static KDE_QUERY_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-fn kwin_query_script_path(query_id: u64) -> String {
-    let uid = unsafe { libc::getuid() };
-    let request_id = Uuid::new_v4().hyphenated().to_string();
-    format!(
-        "/tmp/kanata-switcher-kwin-query-{}-{}-{}.js",
-        uid, query_id, request_id
-    )
-}
-
-fn kwin_query_probe_script_path(probe_id: u64) -> String {
-    let uid = unsafe { libc::getuid() };
-    let request_id = Uuid::new_v4().hyphenated().to_string();
-    format!(
-        "/tmp/kanata-switcher-kwin-query-probe-{}-{}-{}.js",
-        uid, probe_id, request_id
-    )
-}
-
-fn kwin_runtime_script_path() -> String {
-    let uid = unsafe { libc::getuid() };
-    let request_id = Uuid::new_v4().hyphenated().to_string();
-    format!("/tmp/kanata-switcher-kwin-{}-{}.js", uid, request_id)
-}
-
-#[derive(Debug)]
-struct KdeFocusQueryService {
-    sender: TokioMutex<Option<oneshot::Sender<WindowInfo>>>,
-}
-
-#[zbus::interface(name = "com.github.kanata.Switcher.KdeQuery")]
-impl KdeFocusQueryService {
-    #[allow(non_snake_case)]
-    async fn Focus(&self, window_class: &str, window_title: &str) {
-        let win = WindowInfo {
-            class: window_class.to_string(),
-            title: window_title.to_string(),
-            is_native_terminal: false,
-        };
-        let mut sender = self.sender.lock().await;
-        if let Some(tx) = sender.take() {
-            let _ = tx.send(win);
-        }
-    }
-}
-
-fn kwin_script_object_path(
-    script_num: i32,
-    is_kde6: bool,
-) -> Result<OwnedObjectPath, Box<dyn std::error::Error + Send + Sync>> {
-    let path = if is_kde6 {
-        format!("/Scripting/Script{}", script_num)
-    } else {
-        format!("/{}", script_num)
-    };
-    let obj_path: OwnedObjectPath = path.as_str().try_into()?;
-    Ok(obj_path)
-}
-
-async fn load_kwin_script(
-    connection: &Connection,
-    script_path: &str,
-    is_kde6: bool,
-    cleanup_existing: bool,
-) -> Result<(OwnedObjectPath, &'static str), Box<dyn std::error::Error + Send + Sync>> {
-    if cleanup_existing {
-        for _ in 0..5 {
-            let result = connection
-                .call_method(
-                    Some("org.kde.KWin"),
-                    "/Scripting",
-                    Some("org.kde.kwin.Scripting"),
-                    "loadScript",
-                    &(&script_path,),
-                )
-                .await;
-
-            if result.is_ok() {
-                break;
-            }
-            tokio::time::sleep(Duration::from_secs(1)).await;
-        }
-
-        let _ = connection
-            .call_method(
-                Some("org.kde.KWin"),
-                "/Scripting",
-                Some("org.kde.kwin.Scripting"),
-                "unloadScript",
-                &(&script_path,),
-            )
-            .await;
-    }
-
-    let load_result = connection
-        .call_method(
-            Some("org.kde.KWin"),
-            "/Scripting",
-            Some("org.kde.kwin.Scripting"),
-            "loadScript",
-            &(&script_path,),
-        )
-        .await?;
-
-    let script_num: i32 = load_result.body().deserialize()?;
-    let obj_path = kwin_script_object_path(script_num, is_kde6)?;
-    Ok((obj_path, "org.kde.kwin.Script"))
-}
-
-fn build_kde_query_script(is_kde6: bool, bus_name: &str, object_path: &str) -> String {
-    let active_window = if is_kde6 {
-        "activeWindow"
-    } else {
-        "activeClient"
-    };
-    format!(
-        r#"function reportFocus(client) {{
-  callDBus(
-    "{bus}",
-    "{path}",
-    "{iface}",
-    "{method}",
-    client ? (client.resourceClass || "") : "",
-    client ? (client.caption || "") : ""
-  );
-}}
-reportFocus(workspace.{active});
-"#,
-        bus = bus_name,
-        path = object_path,
-        iface = KDE_QUERY_INTERFACE,
-        method = KDE_QUERY_METHOD,
-        active = active_window
-    )
-}
-
-pub(crate) async fn query_kde_focus(
-    connection: &Connection,
-    is_kde6: bool,
-) -> Result<WindowInfo, Box<dyn std::error::Error + Send + Sync>> {
-    let unique_name = connection
-        .unique_name()
-        .ok_or("KDE focus query requires a unique DBus name")?;
-    let query_id = KDE_QUERY_COUNTER.fetch_add(1, Ordering::SeqCst);
-    let query_path = format!("/com/github/kanata/Switcher/KdeQuery{}", query_id);
-    let (sender, receiver) = oneshot::channel();
-    let service = KdeFocusQueryService {
-        sender: TokioMutex::new(Some(sender)),
-    };
-    connection
-        .object_server()
-        .at(query_path.as_str(), service)
-        .await?;
-
-    let script_path = kwin_query_script_path(query_id);
-    let script = build_kde_query_script(is_kde6, unique_name.as_str(), query_path.as_str());
-    fs::write(&script_path, script)?;
-
-    let (script_obj_path, script_interface) =
-        load_kwin_script(connection, &script_path, is_kde6, false).await?;
-
-    let _kwin_query_guard = KwinScriptGuard::new(
-        connection.clone(),
-        tokio::runtime::Handle::current(),
-        script_path.clone(),
-        script_obj_path.clone(),
-        script_interface,
-    );
-
-    connection
-        .call_method(
-            Some("org.kde.KWin"),
-            script_obj_path,
-            Some(script_interface),
-            "run",
-            &(),
-        )
-        .await?;
-
-    let win = tokio::time::timeout(Duration::from_secs(5), receiver)
-        .await
-        .map_err(|_| "Timed out waiting for KDE focus callback")?
-        .map_err(|_| "KDE focus callback sender dropped")?;
-
-    Ok(win)
-}
-
-pub(crate) async fn query_gnome_focus(
-    connection: &Connection,
-) -> Result<WindowInfo, Box<dyn std::error::Error + Send + Sync>> {
-    let reply = connection
-        .call_method(
-            Some(GNOME_SHELL_BUS_NAME),
-            GNOME_FOCUS_OBJECT_PATH,
-            Some(GNOME_FOCUS_INTERFACE),
-            GNOME_FOCUS_METHOD,
-            &(),
-        )
-        .await?;
-    let (class, title): (String, String) = reply.body().deserialize()?;
-    Ok(WindowInfo {
-        class,
-        title,
-        is_native_terminal: false,
-    })
-}
-
-pub(crate) async fn query_focus_for_env(
-    env: Environment,
-    connection: Option<&Connection>,
-    is_kde6: bool,
-) -> Result<WindowInfo, Box<dyn std::error::Error + Send + Sync>> {
-    match env {
-        Environment::Gnome => {
-            let conn = connection.expect("GNOME focus query requires session connection");
-            query_gnome_focus(conn).await
-        }
-        Environment::Kde => {
-            let conn = connection.expect("KDE focus query requires session connection");
-            query_kde_focus(conn, is_kde6).await
-        }
-        Environment::Wayland => {
-            let display_override = resolve_display_override_for_environment(env, "Focus").await;
-            tokio::task::block_in_place(move || {
-                query_wayland_active_window(display_override.as_deref())
-            })
-        }
-        Environment::X11 => {
-            let display_override = resolve_display_override_for_environment(env, "Focus").await;
-            tokio::task::block_in_place(move || {
-                query_x11_active_window(display_override.as_deref())
-            })
-        }
-        Environment::LinuxConsoleWithLogind => Ok(native_terminal_window()),
-        Environment::Unknown => Ok(WindowInfo::default()),
-    }
-}
-
-pub(crate) async fn apply_focus_for_env(
-    env: Environment,
-    connection: Option<&Connection>,
-    is_kde6: bool,
-    handler: &Arc<Mutex<FocusHandler>>,
-    status_broadcaster: &StatusBroadcaster,
-    pause_broadcaster: &PauseBroadcaster,
-    kanata: &KanataClient,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let win = query_focus_for_env(env, connection, is_kde6).await?;
-    let default_layer = kanata.default_layer().await.unwrap_or_default();
-    if let Some(actions) = handle_focus_event(
-        handler,
-        status_broadcaster,
-        pause_broadcaster,
-        &win,
-        kanata,
-        &default_layer,
-    )
-    .await
-    {
-        execute_focus_actions(kanata, actions).await;
-    }
-    Ok(())
-}
 
 pub(crate) fn map_run_outcome_to_backend_exit(outcome: RunOutcome) -> BackendExit {
     match outcome {
@@ -2221,188 +1957,6 @@ async fn resolve_runtime_unpause_context(
     }
 }
 
-async fn resolve_kde_runtime_query_mode_with_retry(
-    connection: &Connection,
-) -> Result<bool, Box<dyn std::error::Error + Send + Sync>> {
-    let mut last_error: Option<Box<dyn std::error::Error + Send + Sync>> = None;
-
-    for attempt in 0..KDE_RUNTIME_QUERY_MODE_MAX_ATTEMPTS {
-        match ensure_kde_scripting_ready(connection).await {
-            Ok(()) => match resolve_kde_runtime_query_mode(connection).await {
-                Ok(is_kde6) => return Ok(is_kde6),
-                Err(error) => {
-                    eprintln!(
-                        "[KDE] Runtime query mode probe attempt {}/{} failed: {}",
-                        attempt + 1,
-                        KDE_RUNTIME_QUERY_MODE_MAX_ATTEMPTS,
-                        error
-                    );
-                    last_error = Some(error);
-                }
-            },
-            Err(error) => {
-                eprintln!(
-                    "[KDE] Scripting readiness check attempt {}/{} failed: {}",
-                    attempt + 1,
-                    KDE_RUNTIME_QUERY_MODE_MAX_ATTEMPTS,
-                    error
-                );
-                last_error = Some(error);
-            }
-        }
-
-        if attempt + 1 < KDE_RUNTIME_QUERY_MODE_MAX_ATTEMPTS {
-            tokio::time::sleep(KDE_RUNTIME_QUERY_MODE_RETRY_DELAY).await;
-        }
-    }
-
-    Err(last_error.unwrap_or_else(|| {
-        std::io::Error::other("[KDE] Runtime query mode probe failed without error").into()
-    }))
-}
-
-async fn ensure_kde_scripting_ready(
-    connection: &Connection,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let dbus = zbus::fdo::DBusProxy::new(connection).await?;
-    let has_owner = dbus
-        .name_has_owner(KDE_KWIN_BUS_NAME.try_into().unwrap())
-        .await?;
-    if !has_owner {
-        return Err(
-            std::io::Error::other(format!("[KDE] {} is not owned", KDE_KWIN_BUS_NAME)).into(),
-        );
-    }
-
-    if dbus
-        .name_has_owner(KDE_KWIN_SCRIPTING_INTERFACE.try_into().unwrap())
-        .await?
-    {
-        return Ok(());
-    }
-
-    let introspection = connection
-        .call_method(
-            Some(KDE_KWIN_BUS_NAME),
-            KDE_KWIN_SCRIPTING_PATH,
-            Some(DBUS_INTROSPECTABLE_INTERFACE),
-            "Introspect",
-            &(),
-        )
-        .await?;
-    let introspection_xml: String = introspection.body().deserialize()?;
-    if !introspection_xml.contains(KDE_KWIN_SCRIPTING_INTERFACE) {
-        return Err(std::io::Error::other(format!(
-            "[KDE] {} is not exported on {}",
-            KDE_KWIN_SCRIPTING_INTERFACE, KDE_KWIN_SCRIPTING_PATH
-        ))
-        .into());
-    }
-
-    Ok(())
-}
-
-async fn resolve_kde_runtime_query_mode(
-    connection: &Connection,
-) -> Result<bool, Box<dyn std::error::Error + Send + Sync>> {
-    let probe_id = KDE_QUERY_COUNTER.fetch_add(1, Ordering::SeqCst);
-    let script_path = kwin_query_probe_script_path(probe_id);
-    fs::write(&script_path, "function kanataSwitcherProbe() {}\n")?;
-
-    let load_result = connection
-        .call_method(
-            Some(KDE_KWIN_BUS_NAME),
-            KDE_KWIN_SCRIPTING_PATH,
-            Some(KDE_KWIN_SCRIPTING_INTERFACE),
-            "loadScript",
-            &(&script_path,),
-        )
-        .await;
-    let load_reply = match load_result {
-        Ok(reply) => reply,
-        Err(error) => {
-            let _ = remove_kwin_probe_script_file(&script_path);
-            return Err(Box::new(error));
-        }
-    };
-
-    let script_num: i32 = match load_reply.body().deserialize() {
-        Ok(script_num) => script_num,
-        Err(error) => {
-            let _ = unload_kwin_script_by_path(connection, &script_path).await;
-            let _ = remove_kwin_probe_script_file(&script_path);
-            return Err(Box::new(error));
-        }
-    };
-
-    let kde6_path = format!("/Scripting/Script{}", script_num);
-    let kde5_path = format!("/{}", script_num);
-    let kde6_path_exists = kwin_object_path_exists(connection, kde6_path.as_str()).await;
-    let kde5_path_exists = kwin_object_path_exists(connection, kde5_path.as_str()).await;
-
-    unload_kwin_script_by_path(connection, &script_path).await?;
-    remove_kwin_probe_script_file(&script_path)?;
-
-    match (kde6_path_exists, kde5_path_exists) {
-        (true, false) => Ok(true),
-        (false, true) => Ok(false),
-        (true, true) => Err(std::io::Error::other(
-            "[KDE] Runtime query mode probe found both KDE5 and KDE6 script paths",
-        )
-        .into()),
-        (false, false) => Err(std::io::Error::other(
-            "[KDE] Runtime query mode probe found no known KWin script path",
-        )
-        .into()),
-    }
-}
-
-async fn kwin_object_path_exists(connection: &Connection, object_path: &str) -> bool {
-    connection
-        .call_method(
-            Some(KDE_KWIN_BUS_NAME),
-            object_path,
-            Some("org.freedesktop.DBus.Introspectable"),
-            "Introspect",
-            &(),
-        )
-        .await
-        .is_ok()
-}
-
-async fn unload_kwin_script_by_path(
-    connection: &Connection,
-    script_path: &str,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    connection
-        .call_method(
-            Some(KDE_KWIN_BUS_NAME),
-            KDE_KWIN_SCRIPTING_PATH,
-            Some(KDE_KWIN_SCRIPTING_INTERFACE),
-            "unloadScript",
-            &(&script_path,),
-        )
-        .await?;
-    Ok(())
-}
-
-fn remove_kwin_probe_script_file(
-    script_path: &str,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    match fs::remove_file(script_path) {
-        Ok(_) => Ok(()),
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(Box::new(error)),
-    }
-}
-
-fn environment_requires_focus_query_connection(env: Environment) -> bool {
-    matches!(
-        env,
-        Environment::Gnome | Environment::Kde | Environment::Wayland | Environment::X11
-    )
-}
-
 struct DbusServiceRegistration {
     _connection: Connection,
     status_signal_task: tokio::task::JoinHandle<()>,
@@ -2535,346 +2089,6 @@ async fn register_dbus_service_with_runtime_environment(
         status_signal_task,
         pause_signal_task,
     })
-}
-
-// === GNOME Backend ===
-
-pub(crate) async fn run_gnome(
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    restart_handle: RestartHandle,
-    pause_broadcaster: PauseBroadcaster,
-    shutdown_handle: ShutdownHandle,
-) -> Result<RunOutcome, Box<dyn std::error::Error + Send + Sync>> {
-    let focus_query_connection = Connection::session().await?;
-    apply_focus_for_env(
-        Environment::Gnome,
-        Some(&focus_query_connection),
-        false,
-        &handler,
-        &status_broadcaster,
-        &pause_broadcaster,
-        &kanata,
-    )
-    .await?;
-
-    let signal_connection = Connection::session().await?;
-    let focus_signal_subscription = subscribe_to_gnome_focus_signal(
-        &signal_connection,
-        kanata.clone(),
-        handler.clone(),
-        status_broadcaster.clone(),
-        pause_broadcaster.clone(),
-    )
-    .await?;
-
-    println!(
-        "[GNOME] Listening for FocusChanged signals from extension at {}",
-        GNOME_FOCUS_OBJECT_PATH
-    );
-    let outcome = wait_for_restart_or_shutdown(&restart_handle, &shutdown_handle).await;
-    drop(focus_signal_subscription);
-    drop(signal_connection);
-    Ok(outcome)
-}
-
-/// Guard owning the GNOME FocusChanged signal subscription task. Dropping the
-/// guard aborts the listener (and releases the match rule when the connection
-/// is dropped).
-struct GnomeFocusSignalSubscription {
-    task: tokio::task::JoinHandle<()>,
-}
-
-impl Drop for GnomeFocusSignalSubscription {
-    fn drop(&mut self) {
-        self.task.abort();
-    }
-}
-
-async fn subscribe_to_gnome_focus_signal(
-    connection: &Connection,
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    pause_broadcaster: PauseBroadcaster,
-) -> Result<GnomeFocusSignalSubscription, Box<dyn std::error::Error + Send + Sync>> {
-    use zbus::MatchRule;
-    let match_rule = MatchRule::builder()
-        .msg_type(zbus::message::Type::Signal)
-        .sender(GNOME_SHELL_BUS_NAME)
-        .map_err(|error| format!("Invalid sender match rule: {}", error))?
-        .interface(GNOME_FOCUS_INTERFACE)
-        .map_err(|error| format!("Invalid interface match rule: {}", error))?
-        .path(GNOME_FOCUS_OBJECT_PATH)
-        .map_err(|error| format!("Invalid path match rule: {}", error))?
-        .member(GNOME_FOCUS_SIGNAL)
-        .map_err(|error| format!("Invalid member match rule: {}", error))?
-        .build();
-
-    let mut stream = zbus::MessageStream::for_match_rule(match_rule, connection, None).await?;
-    let task = tokio::spawn(async move {
-        while let Some(message) = stream.next().await {
-            let message = match message {
-                Ok(message) => message,
-                Err(error) => {
-                    eprintln!("[GNOME] FocusChanged signal error: {}", error);
-                    continue;
-                }
-            };
-            let (window_class, window_title): (String, String) =
-                match message.body().deserialize() {
-                    Ok(payload) => payload,
-                    Err(error) => {
-                        eprintln!("[GNOME] FocusChanged decode error: {}", error);
-                        continue;
-                    }
-                };
-            let win = WindowInfo {
-                class: window_class,
-                title: window_title,
-                is_native_terminal: false,
-            };
-            let default_layer = kanata.default_layer().await.unwrap_or_default();
-            if let Some(actions) = handle_focus_event(
-                &handler,
-                &status_broadcaster,
-                &pause_broadcaster,
-                &win,
-                &kanata,
-                &default_layer,
-            )
-            .await
-            {
-                execute_focus_actions(&kanata, actions).await;
-            }
-        }
-    });
-    Ok(GnomeFocusSignalSubscription { task })
-}
-
-// === KDE Backend ===
-
-#[derive(Debug)]
-struct KwinScriptGuard {
-    connection: Connection,
-    runtime_handle: tokio::runtime::Handle,
-    script_path: String,
-    script_obj_path: OwnedObjectPath,
-    script_interface: String,
-}
-
-impl KwinScriptGuard {
-    fn new(
-        connection: Connection,
-        runtime_handle: tokio::runtime::Handle,
-        script_path: String,
-        script_obj_path: OwnedObjectPath,
-        script_interface: &str,
-    ) -> Self {
-        Self {
-            connection,
-            runtime_handle,
-            script_path,
-            script_obj_path,
-            script_interface: script_interface.to_string(),
-        }
-    }
-}
-
-impl Drop for KwinScriptGuard {
-    fn drop(&mut self) {
-        let connection = self.connection.clone();
-        let runtime_handle = self.runtime_handle.clone();
-        let script_path = self.script_path.clone();
-        let script_obj_path = self.script_obj_path.clone();
-        let script_interface = self.script_interface.clone();
-
-        let cleanup = async move {
-            let stop_result = connection
-                .call_method(
-                    Some("org.kde.KWin"),
-                    script_obj_path.clone(),
-                    Some(script_interface.as_str()),
-                    "stop",
-                    &(),
-                )
-                .await;
-            if let Err(error) = stop_result {
-                panic!("[KDE] Failed to stop KWin script: {}", error);
-            }
-
-            let unload_result = connection
-                .call_method(
-                    Some("org.kde.KWin"),
-                    "/Scripting",
-                    Some("org.kde.kwin.Scripting"),
-                    "unloadScript",
-                    &(&script_path,),
-                )
-                .await;
-            if let Err(error) = unload_result {
-                panic!("[KDE] Failed to unload KWin script: {}", error);
-            }
-        };
-
-        if tokio::runtime::Handle::try_current().is_ok() {
-            tokio::task::block_in_place(|| {
-                runtime_handle.block_on(cleanup);
-            });
-        } else {
-            runtime_handle.block_on(cleanup);
-        }
-
-        if let Err(error) = fs::remove_file(&self.script_path) {
-            if error.kind() != std::io::ErrorKind::NotFound {
-                panic!("[KDE] Failed to remove KWin script file: {}", error);
-            }
-        }
-    }
-}
-
-/// Build the KWin focus-push script body. Targets the per-instance daemon
-/// bus name so multiple daemons coexist on KDE with isolated push channels.
-fn build_kde_focus_push_script(bus_name: &str, api: &str, active_window: &str) -> String {
-    format!(
-        r#"function notifyFocus(client) {{
-  callDBus(
-    "{bus}",
-    "{path}",
-    "{iface}",
-    "WindowFocus",
-    client ? (client.resourceClass || "") : "",
-    client ? (client.caption || "") : ""
-  );
-}}
-workspace.{api}.connect(notifyFocus);
-notifyFocus(workspace.{active});
-"#,
-        bus = bus_name,
-        path = DBUS_PATH,
-        iface = DBUS_INTERFACE,
-        api = api,
-        active = active_window
-    )
-}
-
-pub(crate) async fn run_kde(
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    restart_handle: RestartHandle,
-    pause_broadcaster: PauseBroadcaster,
-    shutdown_handle: ShutdownHandle,
-    effective_name: String,
-) -> Result<RunOutcome, Box<dyn std::error::Error + Send + Sync>> {
-    let connection = Connection::session().await?;
-    let focus_query_connection = Connection::session().await?;
-    let runtime_handle = tokio::runtime::Handle::current();
-    let is_kde6 = resolve_kde_runtime_query_mode_with_retry(&focus_query_connection).await?;
-
-    apply_focus_for_env(
-        Environment::Kde,
-        Some(&focus_query_connection),
-        is_kde6,
-        &handler,
-        &status_broadcaster,
-        &pause_broadcaster,
-        &kanata,
-    )
-    .await?;
-
-    // Inject KWin script (DBus service is ready to receive calls)
-    let api = if is_kde6 {
-        "windowActivated"
-    } else {
-        "clientActivated"
-    };
-    let active_window = if is_kde6 {
-        "activeWindow"
-    } else {
-        "activeClient"
-    };
-    let kwin_script = build_kde_focus_push_script(&effective_name, api, active_window);
-
-    let script_path = kwin_runtime_script_path();
-    fs::write(&script_path, &kwin_script)?;
-
-    for _ in 0..5 {
-        let result = connection
-            .call_method(
-                Some("org.kde.KWin"),
-                KDE_KWIN_SCRIPTING_PATH,
-                Some(KDE_KWIN_SCRIPTING_INTERFACE),
-                "loadScript",
-                &(&script_path,),
-            )
-            .await;
-
-        if result.is_ok() {
-            break;
-        }
-        tokio::time::sleep(Duration::from_secs(1)).await;
-    }
-
-    let _ = connection
-        .call_method(
-            Some("org.kde.KWin"),
-            KDE_KWIN_SCRIPTING_PATH,
-            Some(KDE_KWIN_SCRIPTING_INTERFACE),
-            "unloadScript",
-            &(&script_path,),
-        )
-        .await;
-
-    let load_result = connection
-        .call_method(
-            Some("org.kde.KWin"),
-            KDE_KWIN_SCRIPTING_PATH,
-            Some(KDE_KWIN_SCRIPTING_INTERFACE),
-            "loadScript",
-            &(&script_path,),
-        )
-        .await?;
-
-    let script_num: i32 = load_result.body().deserialize()?;
-
-    let script_obj_path_str = if is_kde6 {
-        format!("/Scripting/Script{}", script_num)
-    } else {
-        format!("/{}", script_num)
-    };
-
-    let script_interface = if is_kde6 {
-        "org.kde.kwin.Script"
-    } else {
-        KDE_KWIN_SCRIPTING_INTERFACE
-    };
-
-    let script_obj_path: OwnedObjectPath = script_obj_path_str.as_str().try_into()?;
-
-    let _kwin_script_guard = KwinScriptGuard::new(
-        connection.clone(),
-        runtime_handle.clone(),
-        script_path.clone(),
-        script_obj_path.clone(),
-        script_interface,
-    );
-
-    connection
-        .call_method(
-            Some("org.kde.KWin"),
-            script_obj_path,
-            Some(script_interface),
-            "run",
-            &(),
-        )
-        .await?;
-
-    println!("[KDE] KWin script injected, listening for window focus events...");
-
-    let outcome = wait_for_restart_or_shutdown(&restart_handle, &shutdown_handle).await;
-    Ok(outcome)
 }
 
 const DBUS_RECONNECT_DELAYS_MS: &[u64] = &[250, 1000, 2000];
